requirements.txt
streamlit
streamlit-autorefresh
pandas
numpy
requests
python-dotenv
